requirements.txt
pyelftools==0.32
pycryptodome==3.20.0